Configuration for Marlin 2.1 Bugfix
Make sure to use the Correct version! It is Marlin 2.1 BUGFIX

This is the most current Marlin version for the Fabrikator after contacting Marlin Team.
This here is a bit better configured Firmware for the Turnigy Fabrikator then the one you can find in Marlin Example Configuration in the official Marlin Repository (I provided them the config). However, here are some values tweaked and some useful functions added like moving from one corner to another for bed leveling.
This firmware assumes that you have the stock Version of the Turnigy Fabrikator without hardware modification. just as after buying.

I will contact Marlin Team again to ask them to add this to the repository of example configurations. When i did that i will update this here in the text.

This is my final version of configuration which should work perfectly, but there will be no updates from my part anymore since i moved to Klipper. The Configuration for Klipper will be uploaded here on my Github too when i am done configurating it as best as I can.

For questions related to this printer you can contact me on Discord PsychoOW#6076

Have fun printing!